Refactor code from ProposalRewardsTable and ProposalRewards using a hook
