+++ components/proposals/ProposalPage/components/ProposalProperties/components/ProposalRewards/ProposalRewards.tsx
@@ -2,29 +2,21 @@ import type { ProposalReviewer } from '@charmverse/core/prisma';
 import { Delete, Edit } from '@mui/icons-material';
 import VisibilityIcon from '@mui/icons-material/Visibility';
 import { Box, Grid, Hidden, IconButton, Stack, Tooltip, Typography } from '@mui/material';
-import { uniqBy } from 'lodash';
-import { useMemo, useState } from 'react';
-import { v4 } from 'uuid';
 
 import { SelectPreviewContainer } from 'components/common/DatabaseEditor/components/properties/TagSelect/TagSelect';
 import { NewDocumentPage } from 'components/common/PageDialog/components/NewDocumentPage';
-import { useNewPage } from 'components/common/PageDialog/hooks/useNewPage';
 import { NewPageDialog } from 'components/common/PageDialog/NewPageDialog';
 import { MilestonePropertiesForm } from 'components/rewards/components/RewardProperties/MilestonePropertiesForm';
 import { RewardAmount } from 'components/rewards/components/RewardStatusBadge';
-import { useNewReward } from 'components/rewards/hooks/useNewReward';
 import { useRewardPage } from 'components/rewards/hooks/useRewardPage';
 import { useRewards } from 'components/rewards/hooks/useRewards';
 import { useRewardsNavigation } from 'components/rewards/hooks/useRewardsNavigation';
-import { useRewardTemplates } from 'components/rewards/hooks/useRewardTemplates';
 import { useCharmRouter } from 'hooks/useCharmRouter';
 import { useSpaceFeatures } from 'hooks/useSpaceFeatures';
 import type { ProposalPendingReward } from 'lib/proposals/interfaces';
-import { getRewardErrors } from 'lib/rewards/getRewardErrors';
-import type { RewardTemplate } from 'lib/rewards/getRewardTemplates';
-import { getRewardType } from 'lib/rewards/getRewardType';
-import type { RewardReviewer } from 'lib/rewards/interfaces';
 import { isTruthy } from 'lib/utils/types';
+
+import { useProposalRewards } from '../../hooks/useProposalRewards';
 
 import { AttachRewardButton } from './AttachRewardButton';
 
@@ -56,12 +48,8 @@ export function ProposalRewards({
   isProposalTemplate
 }: Props) {
   useRewardsNavigation(rewardQueryKey);
-  const { isDirty, clearNewPage, openNewPage, newPageValues, updateNewPageValues } = useNewPage();
-  const { clearRewardValues, contentUpdated, rewardValues, setRewardValues, isSavingReward } = useNewReward();
-  const [currentPendingId, setCurrentPendingId] = useState<null | string>(null);
   const { getRewardPage } = useRewardPage();
   const { rewards: allRewards } = useRewards();
-  const { templates } = useRewardTemplates({ load: !!requiredTemplateId });
 
   const { getFeatureTitle } = useSpaceFeatures();
   const {
@@ -72,70 +60,27 @@ export function ProposalRewards({
   const rewards = rewardIds?.map((rId) => allRewards?.find((r) => r.id === rId)).filter(isTruthy) || [];
   const canCreatePendingRewards = !readOnly && !rewardIds?.length;
 
-  function closeDialog() {
-    clearRewardValues();
-    clearNewPage();
-    setCurrentPendingId(null);
-  }
-
-  async function saveForm() {
-    if (newPageValues) {
-      onSave({ reward: rewardValues, page: newPageValues, draftId: currentPendingId || '' });
-      closeDialog();
-    }
-  }
-
-  const rewardReviewers = useMemo(() => {
-    return uniqBy(
-      reviewers
-        .map((reviewer) =>
-          reviewer.roleId
-            ? { group: 'role', id: reviewer.roleId }
-            : reviewer.userId
-            ? { group: 'user', id: reviewer.userId }
-            : null
-        )
-        .filter(isTruthy) as RewardReviewer[],
-      'id'
-    );
-  }, [reviewers]);
-
-  function getTemplateAppliedReward(template?: RewardTemplate) {
-    const rewardType = template ? getRewardType(template.reward) : 'none';
-    return {
-      ...template?.reward,
-      reviewers: rewardReviewers,
-      assignedSubmitters,
-      // Converting reward values to be of assigned workflow
-      approveSubmitters: false,
-      allowMultipleApplications: false,
-      allowedSubmitterRoles: [],
-      maxSubmissions: null,
-      rewardType
-    };
-  }
-
-  function createNewReward() {
-    clearRewardValues();
-    const template = templates?.find((t) => t.page.id === requiredTemplateId);
-    setRewardValues(getTemplateAppliedReward(template), { skipDirty: true });
-
-    openNewPage({
-      ...template?.page,
-      content: template?.page.content as any,
-      templateId: requiredTemplateId || undefined,
-      title: undefined,
-      type: 'bounty'
-    });
-    // set a new draftId
-    setCurrentPendingId(v4());
-  }
-
-  function showReward({ reward, page, draftId }: ProposalPendingReward) {
-    setRewardValues(reward);
-    openNewPage(page || undefined);
-    setCurrentPendingId(draftId);
-  }
+  const {
+    createNewReward,
+    closeDialog,
+    newRewardErrors,
+    saveForm,
+    selectTemplate,
+    contentUpdated,
+    isDirty,
+    isSavingReward,
+    newPageValues,
+    showReward,
+    setRewardValues,
+    updateNewPageValues,
+    rewardValues
+  } = useProposalRewards({
+    assignedSubmitters,
+    isProposalTemplate,
+    onSave,
+    reviewers,
+    requiredTemplateId
+  });
 
   function openReward(rewardId: string | null) {
     if (!rewardId) return;
@@ -150,32 +95,6 @@ export function ProposalRewards({
 
     updateURLQuery({ [rewardQueryKey]: pageIdToOpen });
   }
-
-  function selectTemplate(template: RewardTemplate | null) {
-    if (template) {
-      setRewardValues(getTemplateAppliedReward(template));
-      updateNewPageValues({
-        ...template.page,
-        content: template.page.content as any,
-        title: undefined,
-        type: 'bounty',
-        templateId: template.page.id
-      });
-    } else {
-      updateNewPageValues({
-        templateId: undefined
-      });
-    }
-  }
-
-  const newRewardErrors = getRewardErrors({
-    page: newPageValues,
-    reward: rewardValues,
-    rewardType: rewardValues.rewardType,
-    isProposalTemplate,
-    isMilestone: true,
-    templateId: newPageValues?.templateId
-  }).join(', ');
 
   if (rewards.length) {
     return (
